refactor: 폴더 구조 리팩토링
- domain/logbat 하위로 Logbat과 AppKey 이동
