+++ sdk/java/logbat-sdk/src/main/java/info/logbat/logback/domain/logbat/Logbat.java
@@ -1,4 +1,4 @@
-package info.logbat.logback;
+package info.logbat.logback.domain.logbat;
 
 public class Logbat {
